Move CloudURLs to chrome types
