+++ packages/core/chrome/core-chrome-browser/index.ts
@@ -36,6 +36,7 @@ export type {
   CloudLinkId,
   CloudLink,
   CloudLinks,
+  CloudURLs,
   SideNavCompProps,
   SideNavComponent,
   SideNavNodeStatus,
